# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GD Progress Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LevelDifficulty = typeof LevelDifficulty[keyof typeof LevelDifficulty];


export const LevelDifficulty = {
  Auto: 'Auto',
  Easy: 'Easy',
  Normal: 'Normal',
  Hard: 'Hard',
  Harder: 'Harder',
  Insane: 'Insane',
  Easy_Demon: 'Easy Demon',
  Medium_Demon: 'Medium Demon',
  Hard_Demon: 'Hard Demon',
  Insane_Demon: 'Insane Demon',
  Extreme_Demon: 'Extreme Demon',
} as const;

export interface Level {
  id: number;
  name: string;
  /** @nullable */
  creator?: string | null;
  difficulty: LevelDifficulty;
  /** @nullable */
  stars?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  bestPercent: number;
  /** @minimum 0 */
  attempts: number;
  isCompleted: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type LevelInputDifficulty = typeof LevelInputDifficulty[keyof typeof LevelInputDifficulty];


export const LevelInputDifficulty = {
  Auto: 'Auto',
  Easy: 'Easy',
  Normal: 'Normal',
  Hard: 'Hard',
  Harder: 'Harder',
  Insane: 'Insane',
  Easy_Demon: 'Easy Demon',
  Medium_Demon: 'Medium Demon',
  Hard_Demon: 'Hard Demon',
  Insane_Demon: 'Insane Demon',
  Extreme_Demon: 'Extreme Demon',
} as const;

export interface LevelInput {
  /** @minLength 1 */
  name: string;
  creator?: string;
  difficulty: LevelInputDifficulty;
  /**
     * @minimum 1
     * @maximum 10
     */
  stars?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  bestPercent?: number;
  /** @minimum 0 */
  attempts?: number;
  notes?: string;
}

export type LevelUpdateDifficulty = typeof LevelUpdateDifficulty[keyof typeof LevelUpdateDifficulty];


export const LevelUpdateDifficulty = {
  Auto: 'Auto',
  Easy: 'Easy',
  Normal: 'Normal',
  Hard: 'Hard',
  Harder: 'Harder',
  Insane: 'Insane',
  Easy_Demon: 'Easy Demon',
  Medium_Demon: 'Medium Demon',
  Hard_Demon: 'Hard Demon',
  Insane_Demon: 'Insane Demon',
  Extreme_Demon: 'Extreme Demon',
} as const;

export interface LevelUpdate {
  /** @minLength 1 */
  name?: string;
  creator?: string;
  difficulty?: LevelUpdateDifficulty;
  /**
     * @minimum 1
     * @maximum 10
     */
  stars?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  bestPercent?: number;
  /** @minimum 0 */
  attempts?: number;
  isCompleted?: boolean;
  notes?: string;
}

export interface DifficultyBreakdown {
  difficulty: string;
  total: number;
  completed: number;
}

export interface Stats {
  totalLevels: number;
  completedLevels: number;
  totalAttempts: number;
  completionRate: number;
  avgBestPercent: number;
  byDifficulty: DifficultyBreakdown[];
}

export interface Session {
  id: number;
  levelId: number;
  /** @minimum 0 */
  attempts: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  bestPercent: number;
  /** @nullable */
  notes?: string | null;
  sessionDate: string;
  createdAt: string;
}

export interface SessionInput {
  /** @minimum 0 */
  attempts: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  bestPercent: number;
  notes?: string;
  sessionDate?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  isOwner: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type GetLevelsParams = {
completed?: boolean;
difficulty?: string;
};

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};