# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GD Progress Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: levels
    description: Level tracking operations
  - name: stats
    description: Statistics operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /levels:
    get:
      operationId: getLevels
      tags: [levels]
      summary: List all tracked levels
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: boolean
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Level"
    post:
      operationId: createLevel
      tags: [levels]
      summary: Add a level to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LevelInput"
      responses:
        "201":
          description: Level created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /levels/{id}:
    get:
      operationId: getLevel
      tags: [levels]
      summary: Get a single level
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Level found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLevel
      tags: [levels]
      summary: Update level progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LevelUpdate"
      responses:
        "200":
          description: Level updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Level"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLevel
      tags: [levels]
      summary: Remove a level from tracking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /levels/{levelId}/sessions:
    get:
      operationId: getLevelSessions
      tags: [levels]
      summary: List all sessions for a level
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createSession
      tags: [levels]
      summary: Log a new practice session for a level
      parameters:
        - name: levelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Session logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Level not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{id}:
    delete:
      operationId: deleteSession
      tags: [levels]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get overall progress statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Level:
      type: object
      required:
        - id
        - name
        - difficulty
        - bestPercent
        - attempts
        - isCompleted
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        creator:
          type: ["string", "null"]
        difficulty:
          type: string
          enum: [Auto, Easy, Normal, Hard, Harder, Insane, "Easy Demon", "Medium Demon", "Hard Demon", "Insane Demon", "Extreme Demon"]
        stars:
          type: ["integer", "null"]
        bestPercent:
          type: integer
          minimum: 0
          maximum: 100
        attempts:
          type: integer
          minimum: 0
        isCompleted:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    LevelInput:
      type: object
      required:
        - name
        - difficulty
      properties:
        name:
          type: string
          minLength: 1
        creator:
          type: string
        difficulty:
          type: string
          enum: [Auto, Easy, Normal, Hard, Harder, Insane, "Easy Demon", "Medium Demon", "Hard Demon", "Insane Demon", "Extreme Demon"]
        stars:
          type: integer
          minimum: 1
          maximum: 10
        bestPercent:
          type: integer
          minimum: 0
          maximum: 100
        attempts:
          type: integer
          minimum: 0
        notes:
          type: string

    LevelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        creator:
          type: string
        difficulty:
          type: string
          enum: [Auto, Easy, Normal, Hard, Harder, Insane, "Easy Demon", "Medium Demon", "Hard Demon", "Insane Demon", "Extreme Demon"]
        stars:
          type: integer
          minimum: 1
          maximum: 10
        bestPercent:
          type: integer
          minimum: 0
          maximum: 100
        attempts:
          type: integer
          minimum: 0
        isCompleted:
          type: boolean
        notes:
          type: string

    Stats:
      type: object
      required:
        - totalLevels
        - completedLevels
        - totalAttempts
        - completionRate
        - byDifficulty
        - avgBestPercent
      properties:
        totalLevels:
          type: integer
        completedLevels:
          type: integer
        totalAttempts:
          type: integer
        completionRate:
          type: number
        avgBestPercent:
          type: number
        byDifficulty:
          type: array
          items:
            $ref: "#/components/schemas/DifficultyBreakdown"

    DifficultyBreakdown:
      type: object
      required:
        - difficulty
        - total
        - completed
      properties:
        difficulty:
          type: string
        total:
          type: integer
        completed:
          type: integer

    Session:
      type: object
      required:
        - id
        - levelId
        - attempts
        - bestPercent
        - sessionDate
        - createdAt
      properties:
        id:
          type: integer
        levelId:
          type: integer
        attempts:
          type: integer
          minimum: 0
        bestPercent:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: ["string", "null"]
        sessionDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    SessionInput:
      type: object
      required:
        - attempts
        - bestPercent
      properties:
        attempts:
          type: integer
          minimum: 0
        bestPercent:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
        sessionDate:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl, isOwner]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        isOwner:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string